Extract signing bits from renderer
