+++ app/cli/internal/action/attestation_push.go
@@ -25,6 +25,7 @@ import (
 	pb "github.com/chainloop-dev/chainloop/app/controlplane/api/controlplane/v1"
 	"github.com/chainloop-dev/chainloop/internal/attestation/crafter"
 	"github.com/chainloop-dev/chainloop/internal/attestation/renderer"
+	clsigner "github.com/chainloop-dev/chainloop/internal/attestation/signer"
 	"github.com/secure-systems-lab/go-securesystemslib/dsse"
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/timestamppb"
@@ -131,9 +132,11 @@ func (action *AttestationPush) Run(ctx context.Context, attestationID string, ru
 	// Indicate that we are done with the attestation
 	action.c.CraftingState.Attestation.FinishedAt = timestamppb.New(time.Now())
 
+	signer := clsigner.NewChainloopSigner(action.keyPath, pb.NewSigningServiceClient(action.CPConnection), action.Logger)
+
 	renderer, err := renderer.NewAttestationRenderer(action.c.CraftingState, action.keyPath, action.cliVersion, action.cliDigest,
 		renderer.WithLogger(action.Logger),
-		renderer.WithCPSigningClient(pb.NewSigningServiceClient(action.CPConnection)))
+		renderer.WithSigner(signer))
 	if err != nil {
 		return nil, err
 	}
